- Update to 132.0.6834.83   * High CVE-2025-0434: Out of bounds memory access in V8   * High CVE-2025-0435: Inappropriate implementation in Navigation   * High CVE-2025-0436: Integer overflow in Skia   * High CVE-2025-0437: Out of bounds read in Metrics   * High CVE-2025-0438: Stack buffer overflow in Tracing   * Medium CVE-2025-0439: Race in Frames   * Medium CVE-2025-0440: Inappropriate implementation in Fullscreen   * Medium CVE-2025-0441: Inappropriate implementation in Fenced   * Medium CVE-2025-0442: Inappropriate implementation in Payments   * Medium CVE-2025-0443: Insufficient data validation in Extensions   * Low CVE-2025-0446: Inappropriate implementation in Extensions   * Low CVE-2025-0447: Inappropriate implementation in Navigation   * Low CVE-2025-0448: Inappropriate implementation in Compositing - Refreshed ppc64 patches - Enabled libtiff for fedora > 41 - Removed useless patches - Fixed some build errors on EL8

+++ chromium-latest.py
@@ -73,10 +73,11 @@ def delete_chromium_files(files):
   full_path = "%s/%s" % (latest_dir, files)
   print('Deleting ' + full_path + ' ', end=' ')
   for filename in glob.glob(full_path):
-    print('Deleting ' + filename + ' ', end=' ')
-    os.remove(filename)
-    print('[DONE]')
-
+    if os.path.isfile(filename):
+      os.remove(filename)
+      print('[DONE]')
+    else:
+      print('[NOT FOUND]')
 
 def check_omahaproxy(channel="stable"):
 
@@ -320,10 +321,14 @@ if __name__ == '__main__':
                  'third_party/node/linux/node-linux-x64',
                  'third_party/rust-toolchain',
                  'third_party/rust-src']
+    junk_files = ['third_party/node/linux/node-linux-x64.tar.gz']
 
     # First, the dirs:
     for directory in junk_dirs:
       delete_chromium_dir(directory)
+    # Remove junk files
+    for file in junk_files:
+      delete_chromium_files(file)
 
   # There has got to be a better, more portable way to do this.
   os.system("find %s -depth -name reference_build -type d -exec rm -rf {} \\;" % latest_dir)
